Grapple Script
It works! Turns off the character controller and player move script when the grapple is initiated, turns them back on when its done. The grapple can also be released by letting go of the mouse button. Now it needs a script to stop it from going crazy fast on grapple and to allow the player to strafe in the air.